move external commit validation to new file, fix feature scoping error
